feat(geometry): conditional compilation for using either f32 or f64

+++ src/object.rs
@@ -5,7 +5,7 @@ use std::{
 
 use wavefront_obj::obj::Primitive;
 
-use crate::{geometry::Intersect, object::triangle::Triangle};
+use crate::{geometry::Intersect, object::triangle::Triangle, Float};
 
 pub mod sphere;
 pub mod triangle;
@@ -26,19 +26,19 @@ pub fn import_from_wavefront_obj_file(path: &str) -> Vec<Box<dyn Intersect>> {
                 if let Primitive::Triangle(a, b, c, ..) = shape.primitive {
                     triangles.push(Box::new(Triangle::from_tuples(
                         (
-                            obj.vertices[a.0].x,
-                            obj.vertices[a.0].y,
-                            obj.vertices[a.0].z,
+                            obj.vertices[a.0].x as Float,
+                            obj.vertices[a.0].y as Float,
+                            obj.vertices[a.0].z as Float,
                         ),
                         (
-                            obj.vertices[b.0].x,
-                            obj.vertices[b.0].y,
-                            obj.vertices[b.0].z,
+                            obj.vertices[b.0].x as Float,
+                            obj.vertices[b.0].y as Float,
+                            obj.vertices[b.0].z as Float,
                         ),
                         (
-                            obj.vertices[c.0].x,
-                            obj.vertices[c.0].y,
-                            obj.vertices[c.0].z,
+                            obj.vertices[c.0].x as Float,
+                            obj.vertices[c.0].y as Float,
+                            obj.vertices[c.0].z as Float,
                         ),
                     )));
                 }
